+++ lib/platform/windows/socket.c
@@ -188,7 +188,7 @@ int socket_get_error(socket_t sock) {
 
 const char *socket_error_string(int error) {
   static char buffer[256];
-  FormatMessageA(FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS, NULL, error,
+  FormatMessageA(FORMAT_MESSAGE_FROM_SYSTEM | FORMAT_MESSAGE_IGNORE_INSERTS, NULL, (DWORD)error,
                  MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT), buffer, sizeof(buffer), NULL);
   return buffer;
 }
